PDF: tag decorations as artifacts and split the crate into modules
