+++ takumi-pdf-wasm/src/lib.rs
@@ -3,33 +3,28 @@
 #![deny(missing_docs)]
 #![deny(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
 
-use std::{
-  collections::HashMap,
-  str::FromStr,
-  sync::{Arc, RwLock},
-};
+mod date;
+mod font;
+mod metadata;
+mod options;
 
-use serde::{Deserialize, Deserializer, de::Error as DeError};
-use serde_bytes::ByteBuf;
+use std::sync::RwLock;
+
 use serde_wasm_bindgen::{from_value, to_value};
 use takumi_bindings_common::{build_font_resource, default_fonts, stylesheet};
 use takumi_core::{
   Fonts,
   layout::node::Node,
-  resources::{
-    font::FontResource,
-    image::{ImageCacheMode, ImageSource as DecodedImage, ResourceCache},
-  },
-  style::{FontFamily, FontStyle as CssFontStyle, FromCssStr, Lang},
-  viewport::Viewport,
+  resources::{font::FontResource, image::ResourceCache},
+  style::{FontFamily, Lang},
 };
-use takumi_pdf::{
-  Attachment, AttachmentRelationship, MeasureOptions, PageMargins, PageOptions, PdfDate,
-  PdfMetadata, PdfOptions, PdfStandard, Tagging,
-};
+use takumi_pdf::{Attachment, MeasureOptions, PdfMetadata, PdfOptions, PdfStandard, Tagging};
 use wasm_bindgen::prelude::*;
 
-fn map_error(error: impl core::fmt::Debug) -> js_sys::Error {
+use crate::font::Font;
+use crate::options::{PdfRenderOptions, decode_images, resolve_geometry};
+
+pub(crate) fn map_error(error: impl core::fmt::Debug) -> js_sys::Error {
   js_sys::Error::new(&format!("{error:?}"))
 }
 
@@ -61,448 +56,6 @@ extern "C" {
   /// JavaScript object representing a measured size.
   #[wasm_bindgen(typescript_type = "MeasuredSize")]
   pub type MeasuredSizeType;
-}
-
-/// Details for loading a custom font.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct FontDetails {
-  name: Option<String>,
-  data: ByteBuf,
-  weight: Option<f64>,
-  style: Option<FontStyle>,
-  subset_of: Option<String>,
-  generic: Option<String>,
-}
-
-/// Font input, either as detailed object or raw buffer.
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum Font {
-  Object(FontDetails),
-  Buffer(ByteBuf),
-}
-
-struct FontStyle(CssFontStyle);
-
-impl<'de> Deserialize<'de> for FontStyle {
-  fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-  where
-    D: Deserializer<'de>,
-  {
-    let value = String::deserialize(deserializer)?;
-    Ok(Self(
-      CssFontStyle::from_css_str(&value).map_err(D::Error::custom)?,
-    ))
-  }
-}
-
-/// An image source with its URL and raw data.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct ImageSource {
-  src: Arc<str>,
-  data: ByteBuf,
-  cache: Option<ImageCacheMode>,
-}
-
-/// Explicit page or viewport dimensions in CSS px.
-#[derive(Deserialize, Clone, Copy)]
-struct Dimensions {
-  width: f32,
-  height: f32,
-}
-
-/// Viewport for single-page output. A missing height sizes the page to the
-/// laid-out content.
-#[derive(Deserialize, Clone, Copy)]
-struct ViewportInput {
-  width: f32,
-  height: Option<f32>,
-}
-
-/// A page size: a preset name (`"a4"`, `"letter"`) or explicit dimensions.
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum SizeInput {
-  Named(String),
-  Dimensions(Dimensions),
-}
-
-/// A page margin: one number for all sides, or per-side values (missing
-/// sides are zero).
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum MarginInput {
-  Uniform(f32),
-  Sides {
-    top: Option<f32>,
-    right: Option<f32>,
-    bottom: Option<f32>,
-    left: Option<f32>,
-  },
-}
-
-fn resolve_page(
-  size: Option<&SizeInput>,
-  landscape: bool,
-  margin: Option<&MarginInput>,
-) -> Result<PageOptions, js_sys::Error> {
-  let mut page = match size {
-    None => PageOptions::A4,
-    Some(SizeInput::Dimensions(dimensions)) => PageOptions {
-      width: dimensions.width,
-      height: dimensions.height,
-      ..PageOptions::A4
-    },
-    Some(SizeInput::Named(name)) => match name.to_ascii_lowercase().as_str() {
-      "a4" => PageOptions::A4,
-      "letter" => PageOptions::LETTER,
-      other => {
-        return Err(js_sys::Error::new(&format!("unknown page size: {other}")));
-      }
-    },
-  };
-
-  if landscape {
-    page = page.landscape();
-  }
-  match margin {
-    None => {}
-    Some(MarginInput::Uniform(value)) => page.margin = PageMargins::uniform(*value),
-    Some(MarginInput::Sides {
-      top,
-      right,
-      bottom,
-      left,
-    }) => {
-      page.margin = PageMargins {
-        top: top.unwrap_or(0.0),
-        right: right.unwrap_or(0.0),
-        bottom: bottom.unwrap_or(0.0),
-        left: left.unwrap_or(0.0),
-      };
-    }
-  }
-  Ok(page)
-}
-
-/// Options for rendering a PDF.
-#[derive(Deserialize, Default)]
-#[serde(rename_all = "camelCase")]
-struct PdfRenderOptions {
-  /// Fixed viewport for single-page output: percentage heights resolve
-  /// against it and overflowing content is clipped. Mutually exclusive with
-  /// the paged fields.
-  viewport: Option<ViewportInput>,
-  /// Page size for paged output. Defaults to A4.
-  size: Option<SizeInput>,
-  /// Swaps the page's width and height, including explicit sizes.
-  landscape: Option<bool>,
-  /// Page margin in CSS px: one number or per-side values.
-  margin: Option<MarginInput>,
-  /// Band repeated at the top of every page. Nodes classed `pageNumber` /
-  /// `totalPages` receive the counters, optionally formatted by a
-  /// supported `@counter-style` name in the same class list.
-  header: Option<Node>,
-  /// Band repeated at the bottom of every page; same class hooks as `header`.
-  footer: Option<Node>,
-  /// Pre-fetched images keyed by URL.
-  images: Option<Vec<ImageSource>>,
-  /// CSS stylesheets to apply before layout.
-  stylesheets: Option<Vec<String>>,
-  /// Per-render font stack: ordered family names used as the fallback chain.
-  font_families: Option<Vec<String>>,
-  /// Default BCP-47 language tag applied to the root.
-  lang: Option<String>,
-  /// Document metadata written to the PDF's info dictionary.
-  metadata: Option<MetadataInput>,
-  /// Generates a PDF outline (bookmarks) from `h1`–`h6` headings.
-  outline: Option<bool>,
-  /// PDF/A conformance level: "2a", "2b", "2u", "3a", "3b", "3u" or "4".
-  pdfa: Option<PdfaInput>,
-  /// Structure-tree emission: `false`, `true` (default) or `"ua1"` to also
-  /// validate against PDF/UA-1.
-  tagged: Option<TaggedInput>,
-  /// Files attached to the document.
-  attachments: Option<Vec<AttachmentInput>>,
-}
-
-/// A file attached to the document.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct AttachmentInput {
-  name: String,
-  data: AttachmentData,
-  mime_type: Option<String>,
-  description: Option<String>,
-  relationship: Option<RelationshipInput>,
-  /// UTC modification date as `YYYY-MM-DD` or `YYYY-MM-DDTHH:MM:SS`.
-  modification_date: Option<String>,
-}
-
-/// Attachment bytes, or a string encoded as UTF-8.
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum AttachmentData {
-  Bytes(ByteBuf),
-  Text(String),
-}
-
-/// AFRelationship names accepted from JS.
-#[derive(Deserialize, Clone, Copy)]
-#[serde(rename_all = "lowercase")]
-enum RelationshipInput {
-  Source,
-  Data,
-  Alternative,
-  Supplement,
-  Unspecified,
-}
-
-impl From<RelationshipInput> for AttachmentRelationship {
-  fn from(relationship: RelationshipInput) -> Self {
-    match relationship {
-      RelationshipInput::Source => Self::Source,
-      RelationshipInput::Data => Self::Data,
-      RelationshipInput::Alternative => Self::Alternative,
-      RelationshipInput::Supplement => Self::Supplement,
-      RelationshipInput::Unspecified => Self::Unspecified,
-    }
-  }
-}
-
-impl TryFrom<AttachmentInput> for Attachment {
-  type Error = js_sys::Error;
-
-  fn try_from(input: AttachmentInput) -> Result<Self, Self::Error> {
-    let modification_date = input
-      .modification_date
-      .as_deref()
-      .map(|value| {
-        parse_date(value).ok_or_else(|| {
-          js_sys::Error::new("invalid modificationDate: expected YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS")
-        })
-      })
-      .transpose()?;
-
-    Ok(Self {
-      name: input.name,
-      data: match input.data {
-        AttachmentData::Bytes(bytes) => bytes.into_vec(),
-        AttachmentData::Text(text) => text.into_bytes(),
-      },
-      mime_type: input.mime_type,
-      description: input.description,
-      relationship: input.relationship.map(Into::into).unwrap_or_default(),
-      modification_date,
-    })
-  }
-}
-
-/// `tagged` values accepted from JS.
-#[derive(Deserialize, Clone, Copy)]
-#[serde(untagged)]
-enum TaggedInput {
-  Enabled(bool),
-  #[serde(with = "ua1_literal")]
-  Ua1,
-}
-
-/// Deserializes the `"ua1"` string literal.
-mod ua1_literal {
-  use serde::{Deserialize, Deserializer};
-
-  pub fn deserialize<'de, D: Deserializer<'de>>(deserializer: D) -> Result<(), D::Error> {
-    let value = String::deserialize(deserializer)?;
-
-    if value == "ua1" {
-      Ok(())
-    } else {
-      Err(serde::de::Error::custom("expected \"ua1\""))
-    }
-  }
-}
-
-impl From<TaggedInput> for Tagging {
-  fn from(tagged: TaggedInput) -> Self {
-    match tagged {
-      TaggedInput::Enabled(false) => Tagging::Off,
-      TaggedInput::Enabled(true) => Tagging::On,
-      TaggedInput::Ua1 => Tagging::Ua1,
-    }
-  }
-}
-
-/// PDF/A conformance level names accepted from JS.
-#[derive(Deserialize, Clone, Copy)]
-enum PdfaInput {
-  #[serde(rename = "2a")]
-  A2a,
-  #[serde(rename = "3a")]
-  A3a,
-  #[serde(rename = "2b")]
-  A2b,
-  #[serde(rename = "2u")]
-  A2u,
-  #[serde(rename = "3b")]
-  A3b,
-  #[serde(rename = "3u")]
-  A3u,
-  #[serde(rename = "4")]
-  A4,
-}
-
-impl From<PdfaInput> for PdfStandard {
-  fn from(pdfa: PdfaInput) -> Self {
-    match pdfa {
-      PdfaInput::A2a => PdfStandard::A2a,
-      PdfaInput::A3a => PdfStandard::A3a,
-      PdfaInput::A2b => PdfStandard::A2b,
-      PdfaInput::A2u => PdfStandard::A2u,
-      PdfaInput::A3b => PdfStandard::A3b,
-      PdfaInput::A3u => PdfStandard::A3u,
-      PdfaInput::A4 => PdfStandard::A4,
-    }
-  }
-}
-
-/// Document metadata fields.
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct MetadataInput {
-  title: Option<String>,
-  description: Option<String>,
-  authors: Option<Vec<String>>,
-  keywords: Option<Vec<String>>,
-  creator: Option<String>,
-  /// UTC creation date as `YYYY-MM-DD` or `YYYY-MM-DDTHH:MM:SS`.
-  creation_date: Option<String>,
-}
-
-fn parse_field<T: FromStr>(part: Option<&str>, width: usize) -> Option<T> {
-  let part = part?;
-
-  if part.len() != width || !part.bytes().all(|byte| byte.is_ascii_digit()) {
-    return None;
-  }
-  part.parse().ok()
-}
-
-fn parse_date(value: &str) -> Option<PdfDate> {
-  let (date, time) = match value.split_once('T') {
-    Some((date, time)) => (date, Some(time.strip_suffix('Z').unwrap_or(time))),
-    None => (value, None),
-  };
-  let mut parts = date.splitn(3, '-');
-  let year: u16 = parse_field(parts.next(), 4)?;
-  let month: u8 = parse_field(parts.next(), 2)?;
-  let day: u8 = parse_field(parts.next(), 2)?;
-  let (hour, minute, second) = match time {
-    Some(time) => {
-      let mut parts = time.splitn(3, ':');
-      (
-        parse_field(parts.next(), 2)?,
-        parse_field(parts.next(), 2)?,
-        parse_field(parts.next(), 2)?,
-      )
-    }
-    None => (0u8, 0u8, 0u8),
-  };
-  let leap = year.is_multiple_of(4) && (!year.is_multiple_of(100) || year.is_multiple_of(400));
-  let days_in_month = match month {
-    1 | 3 | 5 | 7 | 8 | 10 | 12 => 31,
-    4 | 6 | 9 | 11 => 30,
-    2 if leap => 29,
-    2 => 28,
-    _ => return None,
-  };
-
-  if day < 1 || day > days_in_month || hour > 23 || minute > 59 || second > 59 {
-    return None;
-  }
-  Some(PdfDate {
-    year,
-    month,
-    day,
-    hour,
-    minute,
-    second,
-  })
-}
-
-impl TryFrom<MetadataInput> for PdfMetadata {
-  type Error = js_sys::Error;
-
-  fn try_from(input: MetadataInput) -> Result<Self, Self::Error> {
-    let creation_date = input
-      .creation_date
-      .as_deref()
-      .map(|value| {
-        parse_date(value).ok_or_else(|| {
-          js_sys::Error::new("invalid creationDate: expected YYYY-MM-DD or YYYY-MM-DDTHH:MM:SS")
-        })
-      })
-      .transpose()?;
-
-    Ok(Self {
-      title: input.title,
-      description: input.description,
-      authors: input.authors.unwrap_or_default(),
-      keywords: input.keywords.unwrap_or_default(),
-      creator: input.creator,
-      creation_date,
-    })
-  }
-}
-
-fn decode_images(
-  cache: &ResourceCache,
-  sources: Option<Vec<ImageSource>>,
-) -> Result<HashMap<Arc<str>, DecodedImage>, js_sys::Error> {
-  let mut images = HashMap::new();
-
-  for source in sources.unwrap_or_default() {
-    let image = cache
-      .get_or_decode(&source.data, source.cache.unwrap_or_default())
-      .map_err(map_error)?;
-
-    images.insert(source.src, image);
-  }
-  Ok(images)
-}
-
-/// Splits the options into single-page viewport or paged geometry, rejecting
-/// a mix of both.
-fn resolve_geometry(
-  options: &PdfRenderOptions,
-) -> Result<(Option<Viewport>, Option<PageOptions>), js_sys::Error> {
-  let paged_field_set = options.size.is_some()
-    || options.landscape.is_some()
-    || options.margin.is_some()
-    || options.header.is_some()
-    || options.footer.is_some();
-
-  match options.viewport {
-    Some(_) if paged_field_set => Err(js_sys::Error::new(
-      "viewport is mutually exclusive with the paged options (size, landscape, margin, header, footer)",
-    )),
-    Some(input) => Ok((
-      Some(Viewport::new((
-        input.width as u32,
-        input.height.map(|height| height as u32),
-      ))),
-      None,
-    )),
-    None => Ok((
-      None,
-      Some(resolve_page(
-        options.size.as_ref(),
-        options.landscape.unwrap_or(false),
-        options.margin.as_ref(),
-      )?),
-    )),
-  }
 }
 
 /// The size returned by [`PdfRenderer::measure`].
@@ -665,26 +218,5 @@ impl PdfRenderer {
       .map_err(map_error)?
       .unchecked_into(),
     )
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::parse_date;
-
-  #[test]
-  fn parse_date_accepts_documented_formats() {
-    assert!(parse_date("2026-08-06").is_some());
-    assert!(parse_date("2026-08-06T01:02:03").is_some());
-    assert!(parse_date("2026-08-06T01:02:03Z").is_some());
-    assert!(parse_date("2028-02-29").is_some());
-  }
-
-  #[test]
-  fn parse_date_rejects_invalid_input() {
-    assert!(parse_date("2026-08-06T01:02:03ZZ").is_none());
-    assert!(parse_date("2026-02-30").is_none());
-    assert!(parse_date("2026-13-01").is_none());
-    assert!(parse_date("26-08-06").is_none());
   }
 }
