+++ takumi-pdf/src/pagination.rs
@@ -1,0 +1,62 @@
+//! Cutting the content column into pages without splitting unsplittable atoms.
+
+/// Unsplittable vertical extents in content coordinates: text lines, images,
+/// and transformed subtrees (which cannot be windowed without distortion).
+pub(crate) type Atom = (f32, f32);
+
+/// Page start offsets for slicing `total` height into windows of `window`
+/// height. Each cut moves up to the top of any atom straddling it, repeated
+/// until no atom straddles (a raised cut can land inside another atom). An
+/// atom taller than the window can never fit a page, so it does not push cuts
+/// at all — matching browsers, where `break-inside: avoid` is dropped for
+/// boxes taller than the fragmentainer.
+pub(crate) fn page_starts(
+  atoms: &mut [Atom],
+  forced: &mut Vec<f32>,
+  total: f32,
+  window: f32,
+) -> Vec<f32> {
+  atoms.sort_by(|a, b| a.0.total_cmp(&b.0));
+  forced.retain(|cut| *cut > 1.0 && *cut < total - 1.0);
+  forced.sort_by(f32::total_cmp);
+
+  let mut starts = vec![0.0_f32];
+  let mut y0 = 0.0_f32;
+
+  loop {
+    let limit = y0 + window;
+
+    if let Some(cut) = forced.iter().copied().find(|cut| *cut > y0 + 1.0)
+      && cut <= limit
+    {
+      starts.push(cut);
+      y0 = cut;
+      continue;
+    }
+    if limit >= total {
+      break;
+    }
+    let mut cut = limit;
+
+    loop {
+      let pushed_up = atoms
+        .iter()
+        .filter(|(top, bottom)| *top < cut && *bottom > cut && bottom - top <= window)
+        .map(|(top, _)| *top)
+        .fold(cut, f32::min);
+
+      if pushed_up >= cut {
+        break;
+      }
+      if pushed_up <= y0 + 1.0 {
+        cut = limit;
+        break;
+      }
+      cut = pushed_up;
+    }
+
+    starts.push(cut);
+    y0 = cut;
+  }
+  starts
+}
